Move code and rename functions
